==> R_Project_PeaScreen17_Microbiome_25/Scripts/06_Heritability_ind_OTU_fungi.Rmd <==
---
title: "Heritability of individual fungal OTUs"
author: "Valentin Gfeller"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
#global knitter option
knitr::opts_chunk$set(fig.align = "center", echo = FALSE)

# load packages
library(knitr); packageVersion("knitr")
library(tidyverse); packageVersion("tidyverse")
library(phyloseq); packageVersion("phyloseq")
library(vegan); packageVersion("vegan")
library(ape); packageVersion("ape")
library(here); packageVersion("here")
library(readxl); packageVersion("readxl")
library(ggpubr); packageVersion("ggpubr")
library(nlme); packageVersion("nlme")
library(lme4); packageVersion("lme4")
library(FactoMineR); packageVersion("FactoMineR")
library(factoextra); packageVersion("factoextra")
library(HLMdiag); packageVersion("HLMdiag")

# functions were written for repeatedly used code (see "/Scripts/Functions.R")
# source functions
source(here("Scripts", "Functions.R"))


# set global option for ggplot2
theme_set(theme_bw(base_size = 12))
```

# Import data
```{r include=FALSE}
load(here("Data", "Interim", "phy.fun_clr.RData"))

```

Exclude genotypes with less than 2 replicates
```{r }
tab_count <- sample_data(phy.fun_clr) %>% data.frame() %>% 
  group_by(genotype) %>%
  summarise( 
    n = sum(!is.na(emergence))) %>% 
  arrange(n)
table(tab_count$n)

tab_keep <- tab_count %>% filter(n > 1) %>% pull(genotype)

phy.fun_clr <- prune_samples(sample_data(phy.fun_clr)$genotype %in% tab_keep, phy.fun_clr)
```


```{r}
# count table
tab_otu_clr <- phy.fun_clr %>% otu_table() %>% t() %>% data.frame()
tab_otu_clr[1:10, 1:5]

# sample table
tab_samp <- phy.fun_clr %>% sample_data() %>% data.frame()
tab_samp[1:10, 1:5]
```


# Calculate heritablity of individual OTUs
Coded with inspiration from Brachi et al 2022 (https://doi.org/10.1073/pnas.2201285119)

## Function to calculate heritability of many OTUs
```{r}
var_ID <- function(mod){
    var_comp <- varcomp.mer(mod)
    H2 <-  var_comp[2]/sum(var_comp)
    return(as.numeric(H2))
}

var_ID_ind <- function(Y, ID, cov = NULL){
    if(is.null(cov)){
        data <- data.frame(Y = Y, ID = ID)
        mod <-  lmer(Y ~ (1|ID), data)} else{
            data <- data.frame(Y = Y, ID = ID, cov = cov)
            mod <-  lmer(Y ~ cov + (1|ID), data)}
    return(var_ID(mod))}

```

## Calculate heritablity on one OTU (Example)
```{r}
## Run lmer on one OTU
data <- data.frame(Y = tab_otu_clr$fOTU1508, ID = tab_samp$genotype)
mod <- lmer(Y ~ tab_samp$rep + (1|ID), data)

var_comp <- varcomp.mer(mod)
H2 <- var_comp[2]/sum(var_comp)
as.numeric(H2)

## Calculate with normal lm
data <- data.frame(Y = tab_otu_clr[ ,"fOTU1508"], ID = tab_samp$genotype)
mod <- lm(Y ~ ID, data)
summary(mod)$adj.r.squared

# Run bootMER on one OTU
data <- data.frame(Y = tab_otu_clr[ ,"fOTU1508"], ID = tab_samp$genotype)
mod <- lmer(Y ~ tab_samp$rep + (1|ID), data)
b <- bootMer(mod, nsim = 999, FUN = var_ID, seed = 5070,
             parallel = "multicore", ncpus = 30)

med <-  apply(b$t, 2, function(x) as.numeric(median(x, na.rm = TRUE)))
CI.lower = apply(b$t, 2, function(x) as.numeric(quantile(x, probs = 0.025, na.rm=TRUE)))
CI.upper = apply(b$t, 2, function(x) as.numeric(quantile(x, probs = 0.975, na.rm=TRUE)))
tibble('CI lower' = CI.lower, Median = med, 'CI upper' = CI.upper)

```


## Calcualte heritability for all OTUs
```{r}
# Include replicate variable as a covariate
h2 <- apply(tab_otu_clr, 2, var_ID_ind, ID = tab_samp$genotype,  cov = tab_samp$rep)    

h2 <- h2 %>% sort() %>% rev()
h2 %>% sort() %>% tail(n = 10) %>% rev()

```


## Function to calculate CI of heritability (for BootMer)
```{r}
var_boot_ind <- function(Y, ID, cov = NULL, nsim = 999){
    if(is.null(cov)){
        data <- data.frame(Y = Y, ID = ID)
        mod  <-  lmer(Y ~ (1|ID), data)
        b    <- bootMer(mod, nsim = nsim, FUN = var_ID, seed = 5070,
                        parallel = "multicore", ncpus = 30)} else{
            data <- data.frame(Y = Y, ID = ID, cov = cov)
            mod  <-  lmer(Y ~ cov + (1|ID), data)
            b    <- bootMer(mod, nsim = nsim, FUN = var_ID, seed = 5070,
                        parallel = "multicore", ncpus = 30)}
    return(b$t)}
```


## Calcualte CI of heritability for all OTUs
```{r}
# # #Include replicate variable as a covariate
# set.seed(5070)
# t.var_boot <- apply(tab_otu_clr, 2, var_boot_ind, ID = tab_samp$genotype,
#                 cov = tab_samp$rep)
# 
# # #Save as this was computationally heavy calculation
# saveRDS(t.var_boot,
#         here("Data", "Interim", "Bootstrap_fun_OTUs.rds"))

t.var_boot <- readRDS(here("Data", "Interim", "Bootstrap_fun_OTUs.rds"))

tab.CI <- data.frame(OTU = colnames(t.var_boot),
                     median_boot = apply(t.var_boot, 2, median, na.rm = TRUE) * 100,
                     CI_low = apply(t.var_boot,2,  quantile, 0.025,  na.rm = TRUE) * 100,
                     CI_high = apply(t.var_boot,2, quantile, 0.975, na.rm = TRUE) * 100) %>% 
  tibble() %>% arrange(desc(median_boot))

```

# Save heritability
```{r}
# Save heritability and CIs of all OTUs
tab.otu_fun_herit <- tab.CI %>% 
  mutate(heritable = case_when(
    CI_low < 0.001 ~ FALSE,
    TRUE           ~ TRUE),
    heritability = median_boot) %>% 
  select(OTU, heritability, CI_low, CI_high, heritable)

write.csv(tab.otu_fun_herit, row.names = FALSE,
          file = here("Data", "Interim", "heritability_otu_fun.csv"))

```


# Visualize heritability
```{r}
tab.otu_fun_herit %>%
  mutate(OTU = factor(OTU, levels = OTU)) %>%
  ggplot(aes(OTU, heritability )) +
  geom_point() +
  geom_errorbar(aes(ymin = CI_low, ymax = CI_high), width = 0.25) +
  scale_y_continuous(limits = c(0, 100), expand = expansion(mult = c(0, 0))) +
  ylab("Heritability %") +
  theme(panel.grid.major = element_blank(), 
        axis.line = element_line(colour = "black"),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 1.8))

tab.CI %>% 
  filter(CI_low > 2) %>% 
  mutate(OTU = factor(OTU, levels = OTU),
         median_boot = median_boot,
         CI_low  = CI_low,
         CI_high = CI_high) %>% 
  ggplot(aes(OTU, median_boot)) +
  geom_point() +
  geom_errorbar(aes(ymin = CI_low, ymax = CI_high), width = 0.25) +
  scale_y_continuous(limits = c(0, 100), expand = expansion(mult = c(0, 0))) +
  ylab("Heritability %") +
  theme(panel.grid.major = element_blank(), 
        axis.line = element_line(colour = "black"),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1, size = 1.8))

```

```{r}

```
