==> R_Project_PeaScreen17_Microbiome_25/Scripts/23_Representative_sequences.Rmd <==
---
title: "Combine composite figures"
author: "Valentin Gfeller"
date: "`r Sys.Date()`"
output: 
  html_document: 
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
#global knitter option
knitr::opts_chunk$set(fig.align = "center", dev = 'svglite', echo = FALSE)

# load packages
library(tidyverse); packageVersion("tidyverse")
library(phyloseq); packageVersion("phyloseq")
library(vegan); packageVersion("vegan")
library(ape); packageVersion("ape")
library(here); packageVersion("here")
library(readxl); packageVersion("readxl")
library(ggpubr); packageVersion("ggpubr")
library(nlme); packageVersion("nlme")
library(lme4); packageVersion("lme4")
library(knitr); packageVersion("knitr")
library(cowplot); packageVersion("cowplot")
library(factoextra); packageVersion("factoextra")
library(FactoMineR); packageVersion("FactoMineR")
library(ggbeeswarm); packageVersion("ggbeeswarm")
library(car); packageVersion("car")
library(emmeans); packageVersion("emmeans")
library(cowplot); packageVersion("cowplot")
library(UpSetR); packageVersion("UpSetR")

# functions were written for repeatedly used code (see "/Scripts/Functions.R")
# source functions
source(here("Scripts", "Functions.R"))

# set global option for ggplot2
theme_set(theme_bw(base_size = 14))
```

# Tidy representative figures
```{r}
library("Biostrings")
```

## Fungi
```{r}
fastaFile_fun <- readDNAStringSet(here("Data", "Fungi",
                                       "p433_run210325_ITS_AGRIBIOME_ZOTU_c97.fa"))
otu_name = names(fastaFile_fun)
sequence = paste(fastaFile_fun)
df_fungi <- tibble(otu_name, sequence)

# Mean length
sum(width(fastaFile_fun)) / length(fastaFile_fun)

# Change names of sequences to match the OTU names in the manuscript
df_fungi$otu_name <- sub("Z", "f", df_fungi$otu_name)

```

## Bacteria
```{r}
fastaFile_bac <- readDNAStringSet(here("Data", "Bacteria", "16S-otu_0.97-rep-seqs.txt"))
seq_name = names(fastaFile_bac)
sequence = paste(fastaFile_bac)
df_bacteria <- tibble(seq_name, sequence)

# Mean length
sum(width(fastaFile_bac)) / length(fastaFile_bac)

# Change names of sequences to match the OTU names in the manuscript
d.otu_names_bac <- read_csv(here("Data", "Bacteria", "bOTU_names_old_to_new.csv"))

df_bacteria <- df_bacteria %>% 
  left_join(d.otu_names_bac, by = c("seq_name" = "original")) %>% 
  dplyr::rename(otu_name = new) %>% 
  select(otu_name, sequence, seq_name)

```

## Combine domains
```{r}
df_fungi
df_bacteria

df_both <- bind_rows(df_fungi, df_bacteria %>% select(-seq_name))
df_both %>% slice_head(n = 10)
slice_sample(df_both, n = 10)

```

# Print dataframe with representative sequences of OTUs of most interest
```{r}
tab.1 <- c("fOTU860", "fOTU1450", "fOTU1557", "bOTU15", "bOTU5", 
           "fOTU1598", "bOTU1", "fOTU22", "bOTU4", "fOTU1517", "fOTU1508")

tab.S2 <- c("fOTU40", "fOTU419", "fOTU831", "fOTU680", "fOTU578", "fOTU335",
            "fOTU329", "fOTU140", "fOTU334", "fOTU137", "fOTU487", "fOTU1020",
            "fOTU60")
otu_names <- c(tab.1, tab.S2)

d.interest <- df_both %>% 
  filter(otu_name %in% otu_names) %>% 
  mutate(number = str_extract_all(otu_name, "\\d+") %>% as.numeric(unlist(.)),
         character = str_extract_all(otu_name, "[a-z]+") %>% as.character(unlist(.))) %>% 
  arrange(character, number) %>% 
  select(-c(number, character))

d.interest

write.csv(d.interest, row.names = FALSE,
          file = here("Data", "Interim", "Representative_sequence.csv"))

```
